+++ users/tests/test_views.py
@@ -1,5 +1,4 @@
 import os
-import boto3
 from PIL import Image
 from io import BytesIO
 from django.urls import reverse
@@ -78,8 +77,11 @@ class TestViews(TestCase):
     self.response = self.client.get(self.profile_url)
     self.assertEqual(self.response.status_code, 200)
     self.assertTemplateUsed(self.response, 'users/profile.htm')
-    s3 = boto3.resource('s3')
-    s3.Object(
-      os.getenv('AWS_STORAGE_BUCKET_NAME'),
-      'profile_pics/'+image.name
-    ).delete()
+
+    if os.getenv("DEBUG") == 'False':
+      import boto3
+      s3 = boto3.resource('s3')
+      s3.Object(
+        os.getenv('AWS_STORAGE_BUCKET_NAME'),
+        'profile_pics/'+image.name
+      ).delete()
